Refactor homepage
