+++ src/components/homepage/Hero.module.css
@@ -1,0 +1,29 @@
+.heroBanner {
+    padding: 4rem 0;
+    text-align: center;
+    position: relative;
+    overflow: hidden;
+  }
+  
+  .githubButtonWrapper {
+    display: flex;
+  }
+  
+  .githubButton {
+    border: none;
+    margin-left: 24px;
+    overflow: hidden;
+  }
+  
+  @media screen and (max-width: 966px) {
+    .heroBanner {
+      padding: 2rem;
+    }
+  }
+  
+  .buttons {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+  }
+
