+++ src/pages/slack.md
@@ -1,6 +1,10 @@
 # Slack
 
-Join the Capact-related discussion on Slack!
+We have a dedicated channel on CNCF Slack workspace. This is the best place to ask any Capact-related question.
+
+In order to join, follow the steps:
 
 1. Create your Slack account on [https://slack.cncf.io](https://slack.cncf.io).
-2. Join the [`#capact`](https://cloud-native.slack.com/archives/C023YTAHKLG) channel!
+1. Join the [`#capact`](https://cloud-native.slack.com/archives/C023YTAHKLG) channel!
+
+See you there!
